Added VisualOdometer.
Using only the key points detected in last two frame to triangulate and
refine the 3D structure.

For some unknown reason the pose of the last image in the unit test has
a much larger error than all the other images, whose results are almost
the same as the ground truth.